requirements.txt
fastapi
uvicorn[standard]
httpx
pillow
pydantic-settings
torch==2.4.0
torchvision==0.19.0
diffusers==0.31.0
transformers
accelerate
safetensors
